requirements.txt
yt-dlp
openai-whisper
torch